+++ src/types.ts
@@ -16,5 +16,3 @@ export type ExcalidrawVertexElement =
   | ExcalidrawRectangleElement
   | ExcalidrawDiamondElement
   | ExcalidrawEllipseElement;
-
-